requirement.txt
asgiref==3.2.7
Django==3.0.4
django-timezone-field==4.0
djangorestframework==3.11.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.1
